requirements.txt
RPi.GPIO>=0.7.1
picamera2>=0.3.12
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
